backend/requirements.txt
-i https://pypi.org/simple
amqp==5.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.11.3'
attrs==23.2.0; python_version >= '3.7'
billiard==4.2.0; python_version >= '3.7'
celery==5.3.6; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cron-descriptor==1.4.3
da-vinci==0.4.0
debugpy==1.8.1; python_version >= '3.8'
django==5.0.1; python_version >= '3.10'
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.2; python_version >= '3.8'
django-thumbnails==0.7.0
django-timezone-field==6.1.0; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.15.1; python_version >= '3.6'
djecrety==1.0.17; python_version >= '3.6'
drf-spectacular==0.27.2; python_version >= '3.7'
flower==2.0.1; python_version >= '3.7'
humanize==4.9.0; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu==5.3.6; python_version >= '3.8'
markdown==3.6; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
prometheus-client==0.20.0; python_version >= '3.8'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
python-crontab==3.0.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.3; python_version >= '3.7'
referencing==0.34.0; python_version >= '3.8'
rpds-py==0.18.0; python_version >= '3.8'
shortuuid==1.0.13; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tornado==6.4; python_version >= '3.8'
typing-extensions==4.10.0; python_version < '3.11'
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
